Add some new config

+++ apply-config.sh
@@ -5,15 +5,24 @@ source /etc/bigbluebutton/bbb-conf/apply-lib.sh
 
 # Latest version of properties
 cat << EOF
-╔══════════════════════════════════════════════════════════════════════════=╗
-║                                                                           ║
-║ Copy the latest version of "bigbluebutton.properties" file to respective  ║
-║ destination with correct permission                                       ║
-║                                                                           ║
-╚═══════════════════════════════════════════════════════════════════════════╝
+╔═════════════════════════════════════════════=╗
+║                                              ║
+║           Start to apply configs...!         ║
+║       This script made for DarsPlus.com      ║
+║ *** Attention: Don't run on your servers *** ║
+╚══════════════════════════════════════════════╝
 EOF
+echo "This script will run in 15 sec. Press Ctrl+C if you want to stop running the script!!!"
+sleep 15
 
-cp /usr/share/bbb-web/WEB-INF/classes/bigbluebutton.properties bigbluebutton.properties.org
+if [[ ! -f ./bigbluebutton.properties.org ]]
+then
+    cp /usr/share/bbb-web/WEB-INF/classes/bigbluebutton.properties bigbluebutton.properties.org
+    echo " - Backup bigbluebutton.properties ------------------------ [Ok]"
+fi
+
+echo " - bigbluebutton.properties backup exist ------------------ [Ok]"
+
 cp bigbluebutton.properties /usr/share/bbb-web/WEB-INF/classes/bigbluebutton.properties
 chmod 444 /usr/share/bbb-web/WEB-INF/classes/bigbluebutton.properties
 sleep 5
@@ -22,8 +31,17 @@ FQDN=$(sed -n -e '/screenshareRtmpServer/ s/.*\= *//p' bigbluebutton.properties.
 SALT=$(sed -n -e '/securitySalt/ s/.*\= *//p' bigbluebutton.properties.org)
 
 sed -i "s,^bigbluebutton.web.serverURL=.*,bigbluebutton.web.serverURL=https://$FQDN,g" /usr/share/bbb-web/WEB-INF/classes/bigbluebutton.properties
+
+sleep 1
+
 sed -i "s,^screenshareRtmpServer=.*,screenshareRtmpServer=$FQDN,g" /usr/share/bbb-web/WEB-INF/classes/bigbluebutton.properties
+
+sleep 1
+
 sed -i "s,^securitySalt=.*,securitySalt=$SALT,g" /usr/share/bbb-web/WEB-INF/classes/bigbluebutton.properties
+
+echo " - Apply change to bigbluebutton.properties --------------- [Ok]"
+sleep 5
 
 # Last version of settings
 HTML5_CONFIG=/usr/share/meteor/bundle/programs/server/assets/app/config/settings.yml
@@ -50,19 +68,18 @@ sleep 1
 chmod 444 /usr/share/meteor/bundle/programs/server/assets/app/config/settings.yml
 chown 995:995 /usr/share/meteor/bundle/programs/server/assets/app/config/settings.yml
 
-cat << EOF
-╔═════════════════════════════════════════════════════=╗
-║                                                      ║
-║       Change HTML5 client & default web page         ║
-║                                                      ║
-╚══════════════════════════════════════════════════════╝
-EOF
+echo " - Apply new seeting to BBB setting.yml ---------------- [Ok]"
+
+sleep 5
 
 sed -i "s:Source Sans Pro:Vazir:g" /usr/share/meteor/bundle/programs/web.browser/head.html
 sed -i '2i<link href="https://cdn.jsdelivr.net/gh/rastikerdar/vazir-font@v26.0.2/dist/font-face.css" rel="stylesheet" type="text/css" />' /usr/share/meteor/bundle/programs/web.browser/head.html
 
-cd /var/www/bigbluebutton-default && rm -rf * && wget https://github.com/smrezvani/bbb-default-page/archive/main.zip && unzip main.zip && rm -rf main.zip && mv bbb-default-page-main/* . && rm -rf bbb-default-page-main
-sleep 2
+cd /var/www/bigbluebutton-default && rm -rf * && wget https://github.com/smrezvani/bbb-default-page/archive/main.zip && unzip -q main.zip && rm -rf main.zip && mv bbb-default-page-main/* . && rm -rf bbb-default-page-main
+
+echo " - Install default page for BBB --------------------------- [Ok]"
+
+sleep 5
 
 cat << EOF
 ╔════════════════════════════════════════════=╗
